requirements.txt
pyvisa
pyvisa-py
msgpack
pyzmq
jsonschema
importlib_metadata
tblib
.